+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - GP2-3828 - WTE_analytics_grid_markets add
 - GP2-3420 - WTE analytics tabs
 - GP2-3827 - WTE analytics product selector
+- GP2-3830 - Analytics - Learn more about product information (EP)
 
 ## [2.00.0](https://github.com/uktrade/great-cms/releases/tag/2.0.0)
 [Full Changelog](https://github.com/uktrade/great-cms/compare/1.13.3...2.0.0)